Add edit options icon and include functionality
New edit options icon has been added to each line that is editable
by the current user.  On click, show an edit and delete option
that are hooked up to the existing events.

Also include a function to hide the edit options once a user clicks
anywhere on the screen.

Also cleans up the styling as it was pretty ugly.